requirements.txt
influxdb-client
paho-mqtt
RPi.GPIO
dht11
modbus-tk
pyserial
